+++ x/bank/wire.go
@@ -9,3 +9,9 @@ func RegisterWire(cdc *wire.Codec) {
 	cdc.RegisterConcrete(MsgSend{}, "cosmos-sdk/Send", nil)
 	cdc.RegisterConcrete(MsgIssue{}, "cosmos-sdk/Issue", nil)
 }
+
+func getCodec() *wire.Codec {
+	cdc := wire.NewCodec()
+	RegisterWire(cdc)
+	return cdc
+}